Update to java 25 (#478)
* Update to java25

* Fix rebase issues, update dependencies, and add aot cache generation

* Update actions java

* In shadow jar use java 25

* echo java_home

* Rollback jlink plugin version

* Add AOTCache tto jvmArgs

* Fix rebase issues

* Fix determinism issues with noProp masks

+++ shared/src/main/java/com/faforever/neroxis/map/placement/SpawnPlacer.java
@@ -27,7 +27,8 @@ public class SpawnPlacer {
                             SymmetrySettings symmetrySettings) {
         map.getLargeExpansionAIMarkers().clear();
         map.getSpawns().clear();
-        BooleanMask spawnMask = new BooleanMask(map.getSize() + 1, random.nextLong(), symmetrySettings).invert();
+        BooleanMask spawnMask = new BooleanMask(map.getSize() + 1, random.nextLong(),
+                                                symmetrySettings).invert();
         spawnMask.fillSides(map.getSize() / spawnCount * 3 / 2, false)
                  .fillCenter(teamSeparation, false)
                  .fillEdge(map.getSize() / 16, false)
@@ -87,9 +88,9 @@ public class SpawnPlacer {
                      .subtract(
                              new BooleanMask(map.getSize() + 1, random.nextLong(), spawnMask.getSymmetrySettings())
                                      .drawSymmetryLines(spawnMask.getSymmetrySettings().teamSymmetry())
-                                     .inflate((float) teamSeparation / spawnMask.getSymmetrySettings()
-                                                                                .spawnSymmetry()
-                                                                                .getNumSymPoints()))
+                                     .inflate(teamSeparation / spawnMask.getSymmetrySettings()
+                                                                        .spawnSymmetry()
+                                                                        .getNumSymPoints()))
                      .fillEdge(map.getSize() / 32, false)
                      .limitToSymmetryRegion();
         Vector2 location = spawnMaskCopy.getRandomPosition();
